+++ build.py
@@ -1,7 +1,13 @@
 from lxml import etree
 
+with open('templates/overlay.svg', 'rb') as svg_in:
+    svg_template = svg_in.read()
+
+if "tspan1042" in svg_template.decode("utf-8"):
+    print("tspan1042")
+
 # load the template from a file
-svg_template = etree.parse('templates/overlay.svg')
+svg_template = etree.fromstring(svg_template)
 namespace = {'s': 'http://www.w3.org/2000/svg'}
 
 
@@ -41,6 +47,9 @@ tpl_set('image', 'country_flag_p2', '{http://www.w3.org/1999/xlink}href', "{{f.f
 tpl_set('image', 'country_flag_p2', 'clip-path', 'url(#circleClip)')
 
 # ratings p1
+tpl_set('g', 'player_1_rating', 'style', '{{f.showRating(database.players.0)}}')
+tpl_text('tspan', 'player_1_unranked', '{{f.unranked(database.players.0)}}')
+
 tpl_text('tspan', 'rank_p1', '{{f.rank(database.players.0)}}')
 tpl_text('tspan', 'rating_p1', '{{database.players.0.rating}}')
 tpl_text('tspan', 'winrate_p1', '{{f.winrate(database.players.0)}}')
@@ -48,7 +57,10 @@ tpl_text('tspan', 'wins_p1', '{{database.players.0.wins}}')
 tpl_text('tspan', 'losses_p1', '{{database.players.0.losses}}')
 
 # ratings p2
-tpl_text('tspan', 'rank_p2', '{{f.rank(database.players.0)}}')
+tpl_set('g', 'player_2_rating', 'style', '{{f.showRating(database.players.1)}}')
+tpl_text('tspan', 'player_2_unranked', '{{f.unranked(database.players.1)}}')
+
+tpl_text('tspan', 'rank_p2', '{{f.rank(database.players.1)}}')
 tpl_text('tspan', 'rating_p2', '{{database.players.1.rating}}')
 tpl_text('tspan', 'winrate_p2', '{{f.winrate(database.players.1)}}')
 tpl_text('tspan', 'wins_p2', '{{database.players.1.wins}}')
@@ -69,7 +81,8 @@ tpl_set('stop', 'color_start_p2', 'style', 'stop-color:{{f.color(database.match.
 tpl_set('stop', 'color_stop_p2', 'style', 'stop-color:{{f.color(database.match.players.1.color)}};stop-opacity:0;')
 
 svg = etree.tostring(svg_template, encoding='utf8', method='xml')
-
+if "tspan1042" in svg.decode("utf-8"):
+    print("tspan1042")
 
 # inject svg-template into html-template
 with open('templates/overlay.html') as html_in: