+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+Overlay:
+- change circle size
+- fix 2nd player rank
+- display unranked and deranked players
+- layout fixes
 
 
 ## [1.0.0] - 2020-06-03